requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1

numpy==2.2.1
faiss-cpu==1.8.0
sentence-transformers==3.3.1
tqdm==4.67.1
requests==2.32.3

google-api-python-client==2.160.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
beautifulsoup4==4.12.3
